+++ app/src/main/java/woowacourse/shopping/GlobalViewModelFactory.kt
@@ -2,6 +2,7 @@ package woowacourse.shopping
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import woowacourse.shopping.di.DependencyInjector
 
 class GlobalViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
