+++ app/src/main/java/woowacourse/shopping/RepositoryModule.kt
@@ -4,6 +4,8 @@ import woowacourse.shopping.data.CartDefaultRepository
 import woowacourse.shopping.data.CartInMemoryRepository
 import woowacourse.shopping.data.CartProductDao
 import woowacourse.shopping.data.ProductDefaultRepository
+import woowacourse.shopping.di.InMemory
+import woowacourse.shopping.di.RoomDB
 import woowacourse.shopping.model.CartRepository
 import woowacourse.shopping.model.ProductRepository
