+++ app/src/main/java/woowacourse/shopping/di/DependencyInjector.kt
@@ -1,6 +1,5 @@
-package woowacourse.shopping
+package woowacourse.shopping.di
 
-import woowacourse.shopping.DependencyInjector.findInstance
 import kotlin.reflect.KClass
 import kotlin.reflect.KFunction
 import kotlin.reflect.KMutableProperty1
