+++ app/src/main/java/woowacourse/shopping/di/InMemory.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.di
 
 import woowacourse.shopping.data.CartInMemoryRepository
