+++ app/src/main/java/woowacourse/shopping/di/Inject.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.di
 
 @Target(AnnotationTarget.PROPERTY)
 annotation class Inject
