+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import woowacourse.shopping.Inject
-import woowacourse.shopping.RoomDB
+import woowacourse.shopping.di.Inject
+import woowacourse.shopping.di.RoomDB
 import woowacourse.shopping.model.CartProduct
 import woowacourse.shopping.model.CartRepository
